+++ Sources/Sentry/PrivateSentrySDKOnly.mm
@@ -379,6 +379,11 @@ static BOOL _framesTrackingMeasurementHybridSDKMode = NO;
         setRedactContainerClass:containerClass];
 }
 
++ (void)setReplayTags:(NSDictionary<NSString *, id> *)tags
+{
+    [[PrivateSentrySDKOnly getReplayIntegration] setReplayTags:tags];
+}
+
 #endif
 
 @end