+++ Sources/Sentry/Profiling/SentryProfilerTestHelpers.m
@@ -21,7 +21,7 @@ sentry_threadSanitizerIsPresent(void)
     return NO;
 }
 
-#    if defined(TEST) || defined(TESTCI) || defined(DEBUG)
+#    if defined(SENTRY_TEST) || defined(SENTRY_TEST_CI) || defined(DEBUG)
 
 void
 sentry_writeProfileFile(NSData *JSONData)
@@ -68,6 +68,6 @@ sentry_writeProfileFile(NSData *JSONData)
         @"Failed to write data to path %@: %@", pathToWrite, error);
 }
 
-#    endif // defined(TEST) || defined(TESTCI) || defined(DEBUG)
+#    endif // defined(SENTRY_TEST) || defined(SENTRY_TEST_CI) || defined(DEBUG)
 
 #endif // SENTRY_TARGET_PROFILING_SUPPORTED