+++ CHANGELOG.md
@@ -5,11 +5,15 @@
 ### Features
 
 - Add protocol for custom screenName for UIViewControllers (#4646)
+- Allow hybrid SDK to set replay options tags information (#4710)
 - Add threshold to always log fatal logs (#4707)
 
 ### Improvements
 
 - Add error logging for invalid `cacheDirectoryPath` (#4693)
+### Internal
+
+- Change macros TEST and TESTCI to SENTRY_TEST and SENTRY_TEST_CI (#4712)
 
 ## 8.43.1-beta.0